+++ src/Services/OnOfficeParameterConst.php
@@ -69,5 +69,7 @@ trait OnOfficeParameterConst
 
     public const ACTION = 'action';
 
+    public const TYPE = 'type';
+
     public const USER = 'user';
 }